requirements.txt
fastapi
uvicorn
boto3
python-dotenv
supabase
pydantic
streamlit
requests